+++ R/ANFIS.R
@@ -16,10 +16,10 @@ NULL
 #' @author Chao Chen
 #' @references
 #' [1] C. Chen, R. John, J. Twycross, and J. M. Garibaldi, “An extended ANFIS architecture and its learning properties for type-1 and interval type-2 models,” in Proceedings IEEE International Conference on Fuzzy Systems, 2016, pp. 602–609. \cr
-#' \url{https://doi.org/10.1109/FUZZ-IEEE.2016.7737742}
+#' \doi{10.1109/FUZZ-IEEE.2016.7737742}
 #'
 #' [2] C. Chen, R. John, J. Twycross, and J. M. Garibaldi, “Type-1 and interval type-2 ANFIS: a comparison,” in Proceedings IEEE International Conference on Fuzzy Systems, 2017, pp. 1–6.  \cr
-#' \url{https://doi.org/10.1109/FUZZ-IEEE.2017.8015555}
+#' \doi{10.1109/FUZZ-IEEE.2017.8015555}
 #' @export
 
 anfis.builder <- function(fis) {
@@ -232,10 +232,10 @@ anfis.eval <- function(anfis, input.stack) {
 #' @author Chao Chen
 #' @references
 #' [1] C. Chen, R. John, J. Twycross, and J. M. Garibaldi, “An extended ANFIS architecture and its learning properties for type-1 and interval type-2 models,” in Proceedings IEEE International Conference on Fuzzy Systems, 2016, pp. 602–609. \cr
-#' \url{https://doi.org/10.1109/FUZZ-IEEE.2016.7737742}
+#' \doi{10.1109/FUZZ-IEEE.2016.7737742}
 #'
 #' [2] C. Chen, R. John, J. Twycross, and J. M. Garibaldi, “Type-1 and interval type-2 ANFIS: a comparison,” in Proceedings IEEE International Conference on Fuzzy Systems, 2017, pp. 1–6.  \cr
-#' \url{https://doi.org/10.1109/FUZZ-IEEE.2017.8015555}
+#' \doi{10.1109/FUZZ-IEEE.2017.8015555}
 #' @export
 
 anfis.optimise <- function(anfis, data.trn, data.chk=NULL, epoch.total=100, stepsize=0.1, rate.inc=1.1, rate.dec=0.9, method=c("gradient", "lse"), err.log=F, online=0, lambda=1, opt.by="err.opt", err.trn.fix=T) {